# deployment/openapi.yaml
openapi: "3.0.1"
info:
    title: Planly API
    version: "1.0.0"
    description: |
        Planly API - Habits, todos, categories and stats.
        Authenticate via POST /auth/login; use the returned JWT in Authorization header for protected endpoints.
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    schemas:
...
        UpdateTodoStatusRequest:
            type: object
            required: [status, date]
            properties:
                date: { type: string, pattern: "^\\d{4}-\\d{2}-\\d{2}$", example: "2024-01-15" }
                status: { type: string, enum: [done, pending, skipped] }
                progressValue: { type: number, minimum: 0 }
                notes: { type: string }
        UpdateTodoNotesRequest:
            type: object
            required: [notes, date]
            properties:
                date: { type: string, pattern: "^\\d{4}-\\d{2}-\\d{2}$", example: "2024-01-15" }
                notes: { type: string }
        TodoListItem:
            type: object
            description: Habit with todo metadata for a specific date
            properties:
                id: { type: string }
                title: { type: string }
                color: { type: string }
                emoji: { type: string }
                unit: { type: string }
                targetValue: { type: string }
                categoryId: { type: string }
                period: { type: string }
                active: { type: boolean }
                status: { type: string, enum: [done, pending, skipped] }
                progressValue: { type: string }
                notes: { type: string }
                completedAt: { type: string, format: date-time, description: ISO timestamp when habit was marked done }
                streak: { type: integer }
                updatedAt: { type: string }
...
        DailySummaryItem:
            type: object
            properties:
                date: { type: string, pattern: "^\\d{4}-\\d{2}-\\d{2}$" }
                total:
                    type: object
                    properties:
                        done: { type: integer }
                        skipped: { type: integer }
                        pending: { type: integer }
                        total: { type: integer }
                categories:
                    type: array
                    items:
                        type: object
                        properties:
                            categoryId: { type: string }
                            done: { type: integer }
                            skipped: { type: integer }
                            pending: { type: integer }
                            total: { type: integer }
...
        Error:
            type: object
            properties:
                message: { type: string }
                code: { type: string }
                field: { type: string }
paths:
...
    /todo/date:
        get:
            summary: Get habits for a specific date
            operationId: getTodoListByDate
            security:
                - bearerAuth: []
            parameters:
                - name: date
                  in: query
                  required: false
                  description: Date in YYYY-MM-DD format. If not provided, uses current date.
                  schema:
                      type: string
                      pattern: "^\\d{4}-\\d{2}-\\d{2}$"
                      example: "2024-01-15"
            responses:
                "200":
                    description: List of habits for the specified date (with todo metadata when applicable)
                    content:
                        application/json:
                            schema:
                                type: array
                                items: { $ref: "#/components/schemas/TodoListItem" }
                "401":
                    description: Unauthorized
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/Error" }
            x-amazon-apigateway-integration:
                type: aws_proxy
                httpMethod: POST
                uri:
                    Fn::Sub: "arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${TodoLambdaFunction.Arn}/invocations"
        options:
            summary: CORS preflight
            operationId: optionsTodoByDate
            responses:
                "200":
                    description: CORS response
                    headers:
                        Access-Control-Allow-Origin:
                            schema:
                                type: string
                        Access-Control-Allow-Methods:
                            schema:
                                type: string
                        Access-Control-Allow-Headers:
                            schema:
                                type: string
            x-amazon-apigateway-integration:
                type: mock
                requestTemplates:
                    application/json: '{"statusCode": 200}'
                responses:
                    default:
                        statusCode: "200"
                        responseParameters:
                            method.response.header.Access-Control-Allow-Origin: "'*'"
                            method.response.header.Access-Control-Allow-Methods: "'GET,POST,PUT,DELETE,OPTIONS'"
                            method.response.header.Access-Control-Allow-Headers: "'Content-Type,Authorization,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'"
    /todo/summary:
        get:
            summary: Get daily summary of todos
            operationId: getDailySummary
            security:
                - bearerAuth: []
            parameters:
                - name: startDate
                  in: query
                  required: true
                  description: Start date in YYYY-MM-DD format
                  schema:
                      type: string
                      pattern: "^\\d{4}-\\d{2}-\\d{2}$"
                      example: "2024-01-01"
                - name: endDate
                  in: query
                  required: true
                  description: End date in YYYY-MM-DD format
                  schema:
                      type: string
                      pattern: "^\\d{4}-\\d{2}-\\d{2}$"
                      example: "2024-01-31"
            responses:
                "200":
                    description: Daily summary of todos
                    content:
                        application/json:
                            schema:
                                type: array
                                items: { $ref: "#/components/schemas/DailySummaryItem" }
                "400":
                    description: Bad request - startDate and endDate are required
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/Error" }
                "401":
                    description: Unauthorized
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/Error" }
            x-amazon-apigateway-integration:
                type: aws_proxy
                httpMethod: POST
                uri:
                    Fn::Sub: "arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${TodoLambdaFunction.Arn}/invocations"
        options:
            summary: CORS preflight
            operationId: optionsTodoSummary
            responses:
                "200":
                    description: CORS response
                    headers:
                        Access-Control-Allow-Origin:
                            schema:
                                type: string
                        Access-Control-Allow-Methods:
                            schema:
                                type: string
                        Access-Control-Allow-Headers:
                            schema:
                                type: string
            x-amazon-apigateway-integration:
                type: mock
                requestTemplates:
                    application/json: '{"statusCode": 200}'
                responses:
                    default:
                        statusCode: "200"
                        responseParameters:
                            method.response.header.Access-Control-Allow-Origin: "'*'"
                            method.response.header.Access-Control-Allow-Methods: "'GET,POST,PUT,DELETE,OPTIONS'"
                            method.response.header.Access-Control-Allow-Headers: "'Content-Type,Authorization,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'"
    /todo/{habitId}:
        patch:
            summary: Update todo status
            operationId: updateTodoStatus
            security:
                - bearerAuth: []
            parameters:
                - name: habitId
                  in: path
                  required: true
                  schema:
                      type: string
            requestBody:
                content:
                    application/json:
                        schema: { $ref: "#/components/schemas/UpdateTodoStatusRequest" }
            responses:
                "204":
                    description: Todo status updated successfully
                "401":
                    description: Unauthorized
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/Error" }
                "404":
                    description: Todo or habit not found
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/Error" }
            x-amazon-apigateway-integration:
                type: aws_proxy
                httpMethod: POST
                uri:
                    Fn::Sub: "arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${TodoLambdaFunction.Arn}/invocations"
        options:
            summary: CORS preflight
            operationId: optionsTodoStatus
            parameters:
                - name: habitId
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: CORS response
                    headers:
                        Access-Control-Allow-Origin:
                            schema:
                                type: string
                        Access-Control-Allow-Methods:
                            schema:
                                type: string
                        Access-Control-Allow-Headers:
                            schema:
                                type: string
            x-amazon-apigateway-integration:
                type: mock
                requestTemplates:
                    application/json: '{"statusCode": 200}'
                responses:
                    default:
                        statusCode: "200"
                        responseParameters:
                            method.response.header.Access-Control-Allow-Origin: "'*'"
                            method.response.header.Access-Control-Allow-Methods: "'GET,POST,PUT,DELETE,OPTIONS,PATCH'"
                            method.response.header.Access-Control-Allow-Headers: "'Content-Type,Authorization,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'"
    /todo/{habitId}/notes:
        patch:
            summary: Update todo notes
            operationId: updateTodoNotes
            security:
                - bearerAuth: []
            parameters:
                - name: habitId
                  in: path
                  required: true
                  schema:
                      type: string
            requestBody:
                content:
                    application/json:
                        schema: { $ref: "#/components/schemas/UpdateTodoNotesRequest" }
            responses:
                "204":
                    description: Todo notes updated successfully
                "401":
                    description: Unauthorized
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/Error" }
                "404":
                    description: Todo or habit not found
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/Error" }
            x-amazon-apigateway-integration:
                type: aws_proxy
                httpMethod: POST
                uri:
                    Fn::Sub: "arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${TodoLambdaFunction.Arn}/invocations"
        options:
            summary: CORS preflight
            operationId: optionsTodoNotes
            parameters:
                - name: habitId
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: CORS response
                    headers:
                        Access-Control-Allow-Origin:
                            schema:
                                type: string
                        Access-Control-Allow-Methods:
                            schema:
                                type: string
                        Access-Control-Allow-Headers:
                            schema:
                                type: string
            x-amazon-apigateway-integration:
                type: mock
                requestTemplates:
                    application/json: '{"statusCode": 200}'
                responses:
                    default:
                        statusCode: "200"
                        responseParameters:
                            method.response.header.Access-Control-Allow-Origin: "'*'"
                            method.response.header.Access-Control-Allow-Methods: "'GET,POST,PUT,DELETE,OPTIONS,PATCH'"
                            method.response.header.Access-Control-Allow-Headers: "'Content-Type,Authorization,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token'"
...